more informative error messaging

+++ services/QuillLMS/app/controllers/cms/schools_controller.rb
@@ -111,7 +111,7 @@ class Cms::SchoolsController < Cms::CmsController
       flash[:success] = "Yay! It worked! 🎉"
       redirect_to cms_school_path(params[:id])
     rescue
-      flash[:error] = "It did't work! 😭😭😭"
+      flash[:error] = "It didn't work! 😭😭😭"
       redirect_to :back
     end
   end
@@ -119,13 +119,17 @@ class Cms::SchoolsController < Cms::CmsController
   def add_existing_user_by_email
     begin
       user = User.find_by!(email: params[:email_address])
+      raise ArgumentError if user.role != 'teacher'
       school = School.find_by!(id: params[:id])
       SchoolsUsers.where(user: user).destroy_all
       SchoolsUsers.create!(user_id: user.id, school_id: school.id)
       flash[:success] = "Yay! It worked! 🎉"
       redirect_to cms_school_path(params[:id])
     rescue ActiveRecord::RecordNotFound
-      flash[:error] = "It did't work! Make sure the email you typed is correct."
+      flash[:error] = "It didn't work! Make sure the email you typed is correct."
+      redirect_to :back
+    rescue ArgumentError
+      flash[:error] = "It didn't work! Make sure the account you entered belogs to a teacher, not staff or student."
       redirect_to :back
     rescue
       flash[:error] = "It didn't work. See a developer about this issue."
